Sync to 2.6.32.9
The official git tree has no tag for 2.6.32.9 so we just
sync the tree with the 2.6.32.9 release tarball.  We are
using 2.6.32.9 as the base because the kernel the Android
Version ID : 1112091502 (Linux Kernel ANDROID_2.6.32 2011-12-09)
that the wm8650 gpl reference tarball is based on is in
turn based on linux 2.6.32.9.

The WonderMedia 8650 GPL reference tarball we are using
is from https://github.com/vir91/wm8650-gpl-reference-kernel

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>